clean up base entities
